Calendar is more modular, Checkpoint before deleting some lines of code and separate components some more 3
