+++ frontend/src/pages/AppointmentCalendar.tsx
@@ -8,7 +8,7 @@ import getDay from 'date-fns/getDay'
 import enUS from 'date-fns/locale/en-US'
 import 'react-big-calendar/lib/css/react-big-calendar.css'
 
-import type { Appointment } from '../../types.ts'
+import type { Appointment } from './types.ts'
 
 
 const locales = {
